tesseract_api/requirements.txt
Django>=3.2,<4.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pillow==10.3.0
requests==2.32.3
urllib3==2.2.1
pytesseract
